Redesigned/Converted the download modal to Wizard
Changes to match mockups in https://tower-mockups.testing.ansible.com/analytics/reports/analytics-reports-view-export-pdf-01/
