+++ src/Components/Toolbar/DownloadButton/constants.tsx
@@ -1,0 +1,12 @@
+export const actions = {
+  RESET_DATA: 'RESET_DATA',
+  SET_DOWNLOAD_TYPE: 'SET_DOWNLOAD_TYPE',
+  SET_SHOW_EXTRA_ROWS: 'SET_SHOW_EXTRA_ROWS',
+  SET_SELECTED_RBAC_GROUPS: 'SET_SELECTED_RBAC_GROUPS',
+  SET_USERS: 'SET_USERS',
+  SET_ADDITIONAL_RECIPIENTS: 'SET_ADDITIONAL_RECIPIENTS',
+  SET_EULA: 'SET_EULA',
+  SET_EMAIL_EXTRA_ROWS: 'SET_EMAIL_EXTRA_ROWS',
+  SET_SUBJECT: 'SET_SUBJECT',
+  SET_BODY: 'SET_BODY',
+};
